pyproject.toml
[project]
name = "promptv"
version = "0.1.2"
description = "A comprehensive developer-first prompt management tool with git-like version control"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "Thompson Wong", email = "thompsonwong@labs21.dev"},
]
keywords = ['prompt', 'management', 'tool', 'git-like', 'version', 'control', 'developer-first']

dependencies = [
    "click>=8.3.1",
    "pydantic>=2.12.5",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "tiktoken>=0.8.0",
    "textual>=0.84.0",
]
[project.urls]
Homepage = "https://github.com/thompson0012/promptv"
Repository = "https://github.com/thompson0012/promptv"

[project.scripts]
promptv = "promptv.cli:cli"


[dependency-groups]
dev = [
    "ipython>=9.7.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"